addressed a few warnings in the filesystem code.

+++ src/common/filesystem/include/fs_filesystem.h
@@ -119,7 +119,7 @@ public:
 
 	static uint32_t LumpNameHash (const char *name);		// [RH] Create hash key from an 8-char name
 
-	int FileLength (int lump) const;
+	ptrdiff_t FileLength (int lump) const;
 	int GetFileFlags (int lump);					// Return the flags for this lump
 	const char* GetFileShortName(int lump) const;
 	const char *GetFileFullName (int lump, bool returnshort = true) const;	// [RH] Returns the lump's full name
